Raised version of module for ssh_key_pair (#16)
* Raised version of module for ssh_key_pair

This module is pointing to version 0.9.0. However 0.9.0 can only be used with Terraform version 0.12. Result is that this module is broken in 0.12.x+ versions, as you can see in the error that i used with a terraform version 0.13.x:
```
Error: Unsupported Terraform Core version

  on .terraform/modules/instance.ssh_key_pair/versions.tf line 2, in terraform:
   2:   required_version = "~> 0.12.0"
```

In later versions of the ssh_key_pair module this is fixed, so i raised this version right up to the latest available.

* Updated README.md

Co-authored-by: actions-bot <58130806+actions-bot@users.noreply.github.com>